requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# OCR and PDF Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1

# Image Processing and Enhancement
opencv-python==4.8.1.78
numpy==1.26.4

# Data Processing and Validation
pydantic==2.5.0
python-dotenv==1.0.0
regex==2023.10.3

# Utilities
tqdm==4.66.1
click==8.1.7

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For better OCR accuracy
easyocr==1.7.0  # Alternative OCR engine

# Logging and Monitoring
loguru==0.7.2